Remove Facebook Like button
Facebook doesn't track only the time we spend on the site: the Like
button sends over our information too wherever it is displayed.

It is our personal opinion how much there is to lose around privacy
with tracking scripts like these. If awareness of this website doesn't
grow because of the presence of the button there is nothing to gain
either.

See:

* http://www.technologyreview.com/news/541351/facebooks-like-buttons-will-soon-track-your-web-browsing-to-target-ads/
* http://edition.cnn.com/2010/TECH/social.media/06/02/cnet.facebook.privacy.like/index.html
* https://www.eff.org/privacybadger

This patch also removes extra new lines from the end of the file.